Apply translate to Oval
