+++ shoes-swt/lib/shoes/swt.rb
@@ -72,6 +72,7 @@ class Shoes
       require 'shoes/swt/arc_painter.rb'
       require 'shoes/swt/arrow_painter.rb'
       require 'shoes/swt/line_painter.rb'
+      require 'shoes/swt/oval_painter.rb'
       require 'shoes/swt/rect_painter.rb'
       require 'shoes/swt/swt_button'
       require 'shoes/swt/check_button'
